# api/openapi.yaml
openapi: 3.1.0
info:
  title: ProfileWeave Local API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >-
    Loopback-only API for local browser profile isolation and process control.
    It is not designed for LAN or public Internet exposure.
servers:
  - url: http://127.0.0.1:3210/api/v1
security: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Read service health and build information
      responses:
        '200':
          description: Healthy service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '403': { $ref: '#/components/responses/Forbidden' }
  /bootstrap:
    get:
      operationId: getBootstrap
      summary: Get the ephemeral control token for same-origin write requests
      responses:
        '200':
          description: Per-process control token; never persist or log it
          content:
            application/json:
              schema:
                type: object
                required: [controlToken]
                properties:
                  controlToken: { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }
  /shutdown:
    post:
      operationId: shutdownApplication
      summary: Gracefully stop the local application and its managed sessions
      description: >-
        Requests process shutdown after the response is sent. This endpoint is
        loopback-only and requires the current process control token.
      security: [{ ControlToken: [] }]
      responses:
        '202':
          description: Shutdown accepted
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, const: shutting_down }
        '403': { $ref: '#/components/responses/Forbidden' }
        '501':
          description: This embedding does not expose application shutdown
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /capabilities:
    get:
      operationId: getCapabilities
      summary: Discover browsers and runtime capabilities
      responses:
        '200':
          description: Current capabilities
          content:
            application/json:
              schema:
                type: object
                required: [provider, browsers, features]
                properties:
                  provider: { $ref: '#/components/schemas/RuntimeProvider' }
                  browsers:
                    type: array
                    items: { $ref: '#/components/schemas/BrowserDescriptor' }
                  features:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureCapability' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /doctor:
    get:
      operationId: getDoctor
      summary: Diagnose the local runtime provider and browser availability
      description: >-
        Returns safe, structured diagnostics. Provider capabilities describe
        observable support and do not imply that a browser is undetectable.
      responses:
        '200':
          description: Current local runtime diagnostics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DoctorReport' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /profiles:
    get:
      operationId: listProfiles
      summary: List profiles
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Profile' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createProfile
      summary: Create a profile
      security: [{ ControlToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileInput' }
      responses:
        '201':
          description: Created profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/ProfileID'
    get:
      operationId: getProfile
      summary: Get a profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProfile
      summary: Replace a profile using optimistic revision control
      security: [{ ControlToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteProfile
      summary: Delete profile metadata and move browser data to application trash
      security: [{ ControlToken: [] }]
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /profiles/{id}/duplicate:
    post:
      operationId: duplicateProfile
      summary: Duplicate profile configuration with a new isolated data directory
      security: [{ ControlToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProfileID'
      responses:
        '201':
          description: Duplicated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /profiles/{id}/validate:
    post:
      operationId: validateProfile
      summary: Evaluate profile coherence without launching
      security: [{ ControlToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProfileID'
      responses:
        '200':
          description: Coherence report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /profiles/{id}/launch:
    post:
      operationId: launchProfile
      summary: Launch the selected local browser
      security: [{ ControlToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProfileID'
      responses:
        '200':
          description: Active session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /profiles/{id}/stop:
    post:
      operationId: stopProfile
      summary: Stop a browser process owned by this service
      security: [{ ControlToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProfileID'
      responses:
        '200':
          description: Stopped session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /sessions:
    get:
      operationId: listSessions
      summary: List current and recent in-memory sessions
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /trash:
    get:
      operationId: listTrash
      summary: List recoverable deleted profiles
      responses:
        '200':
          description: Recycle bin entries, newest deletion first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/TrashItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /trash/{id}:
    parameters:
      - $ref: '#/components/parameters/ProfileID'
    delete:
      operationId: purgeTrash
      summary: Permanently remove a trashed profile and its browser data
      security: [{ ControlToken: [] }]
      responses:
        '204': { description: Permanently deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/TrashNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /trash/{id}/restore:
    post:
      operationId: restoreTrash
      summary: Restore profile metadata and browser data from the recycle bin
      security: [{ ControlToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProfileID'
      responses:
        '200':
          description: Restored profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/TrashNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    ControlToken:
      type: apiKey
      in: header
      name: X-ProfileWeave-Token
  parameters:
    ProfileID:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^p_[a-f0-9]{32}$'
  responses:
    BadRequest:
      description: Request body is invalid JSON or contains unknown fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Host, Origin, Fetch Metadata, or control token rejected
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Profile not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TrashNotFound:
      description: Trashed profile not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Revision or runtime state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Invalid or incoherent profile
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Health:
      type: object
      required: [status, product, version, commit, date]
      properties:
        status: { type: string, const: ok }
        product: { type: string, const: ProfileWeave }
        version: { type: string }
        commit: { type: string }
        date: { type: string }
    Browser:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [auto, chrome, edge, brave, chromium, custom] }
        customPath: { type: string }
    Screen:
      type: object
      required: [width, height, dpr]
      properties:
        width: { type: integer }
        height: { type: integer }
        dpr: { type: number }
    Fingerprint:
      type: object
      required: [os, uaMode, locale, languages, timezone, screen, cpuCores, memoryGB, webrtcPolicy]
      properties:
        os: { type: string }
        uaMode: { type: string, enum: [native, custom] }
        userAgent: { type: string }
        locale: { type: string }
        languages:
          type: array
          items: { type: string }
        timezone: { type: string }
        screen: { $ref: '#/components/schemas/Screen' }
        cpuCores: { type: integer }
        memoryGB: { type: integer }
        webrtcPolicy: { type: string }
    Proxy:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [direct, http, socks5] }
        host: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535 }
    ProfileInput:
      type: object
      additionalProperties: false
      required: [name, tags, startURL, browser, fingerprint, proxy]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        notes: { type: string, maxLength: 2000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 32 }
        startURL: { type: string }
        browser: { $ref: '#/components/schemas/Browser' }
        fingerprint: { $ref: '#/components/schemas/Fingerprint' }
        proxy: { $ref: '#/components/schemas/Proxy' }
    ProfileUpdate:
      type: object
      additionalProperties: false
      required: [revision, name, tags, startURL, browser, fingerprint, proxy]
      properties:
        id:
          type: string
          pattern: '^p_[a-f0-9]{32}$'
          description: Optional in the body; when present it must match the route ID
        revision: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 100 }
        notes: { type: string, maxLength: 2000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 32 }
        startURL: { type: string }
        browser: { $ref: '#/components/schemas/Browser' }
        fingerprint: { $ref: '#/components/schemas/Fingerprint' }
        proxy: { $ref: '#/components/schemas/Proxy' }
    Profile:
      type: object
      additionalProperties: false
      required: [id, revision, createdAt, updatedAt, name, tags, startURL, browser, fingerprint, proxy]
      properties:
        id:
          type: string
          pattern: '^p_[a-f0-9]{32}$'
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        name: { type: string, minLength: 1, maxLength: 100 }
        notes: { type: string, maxLength: 2000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 32 }
        startURL: { type: string }
        browser: { $ref: '#/components/schemas/Browser' }
        fingerprint: { $ref: '#/components/schemas/Fingerprint' }
        proxy: { $ref: '#/components/schemas/Proxy' }
    TrashItem:
      type: object
      additionalProperties: false
      required: [profile, deletedAt, hasBrowserData]
      properties:
        profile: { $ref: '#/components/schemas/Profile' }
        deletedAt: { type: string, format: date-time }
        hasBrowserData: { type: boolean }
    Session:
      type: object
      required: [profileId, status]
      properties:
        profileId: { type: string }
        status: { type: string, enum: [starting, running, stopping, stopped, failed] }
        pid: { type: integer }
        startedAt: { type: string, format: date-time }
        stoppedAt: { type: string, format: date-time }
        lastError: { type: string }
    BrowserDescriptor:
      type: object
      required: [id, name, available]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        available: { type: boolean }
    FeatureCapability:
      type: object
      required: [key, label, status]
      properties:
        key: { type: string }
        label: { type: string }
        status: { type: string, enum: [applied, partial, unsupported] }
        detail: { type: string }
    ProviderCapability:
      type: object
      required: [id, name, status, detail]
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string, enum: [applied, partial, unsupported] }
        detail: { type: string }
    RuntimeProvider:
      type: object
      required: [id, name, description, source, license, versionManagement, capabilities]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        source: { type: string }
        license: { type: string }
        versionManagement: { type: string }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/ProviderCapability' }
    DoctorIssue:
      type: object
      required: [code, severity, message]
      properties:
        code: { type: string }
        severity: { type: string, enum: [error, warning, info] }
        message: { type: string }
        suggestion: { type: string }
    DoctorReport:
      type: object
      required: [provider, healthy, inspectedBrowsers, availableBrowsers, activeSessions, browsers, issues]
      properties:
        provider: { $ref: '#/components/schemas/RuntimeProvider' }
        healthy: { type: boolean }
        inspectedBrowsers:
          type: integer
          minimum: 0
          description: Number of supported browser descriptors inspected on this host
        availableBrowsers: { type: integer, minimum: 0 }
        activeSessions: { type: integer, minimum: 0 }
        browsers:
          type: array
          items: { $ref: '#/components/schemas/BrowserDescriptor' }
        issues:
          type: array
          items: { $ref: '#/components/schemas/DoctorIssue' }
    Issue:
      type: object
      required: [severity, field, code, message]
      properties:
        severity: { type: string, enum: [error, warning, info] }
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Report:
      type: object
      required: [score, issues]
      properties:
        score: { type: integer, minimum: 0, maximum: 100 }
        issues:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }